firmware: FreeType → TinyTTF, CJK rendering fix + UI refinements
- Roll back e964931 firmware font changes (headers, .ttf, font_fs)
- Recreate noto_emoji.h / cjk_font.h with correct extern declarations
- Switch from FreeType to TinyTTF (lv_tiny_ttf_create_data):
  font data copied to PSRAM, no flash-cache msync issues
- Add CJK font sizes 32/36px; bump session page fonts
  (NAME 24→32, BODY 16→24, SMALL 14→20)
- Notification modal: close on blank-area tap only, not content tap
- Fix session page back-arrow font (montserrat for LV_SYMBOL_UP)
- sdkconfig: disable FreeType, enable TinyTTF
- bridge: serial transport, encoder, frontend SerialTab, config
- Fix esp_lvgl_port managed component dependency (esp_lcd_touch)

+++ firmware/main/assets/font/noto_emoji.h
@@ -1,4 +1,4 @@
-/* noto_emoji.h — Noto Emoji Regular TTF embedded as C array */
+/* noto_emoji.h — Noto Emoji Regular TTF embedded as C array (declarations) */
 #ifndef NOTO_EMOJI_H
 #define NOTO_EMOJI_H
 
@@ -10,7 +10,7 @@ extern "C" {
 #endif
 
 extern const unsigned char emoji_ttf_data[];
-extern const size_t  emoji_ttf_len;
+extern const unsigned int  emoji_ttf_len;
 
 #ifdef __cplusplus
 }
